Refactor marketplace MCP boundaries
